build_etf_model/requirements.txt
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
google-api-core==1.22.4
google-auth==1.22.1
google-cloud==0.34.0
google-cloud-automl==2.0.0
google-cloud-bigquery==2.1.0
google-cloud-bigquery-datatransfer==2.1.0
google-cloud-core==1.4.3
google-cloud-pubsub==2.1.0
google-cloud-storage==1.31.2
google-crc32c==1.0.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.32.0
pandas==1.1.2
pathspec==0.8.0
proto-plus==1.10.2
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.9.27
requests==2.24.0
rsa==4.6
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.3
yarl==1.6.0
yfinance==0.1.55
zipp==3.2.0
